fix: rename base-cmd to match other commands
