+++ lib/arborist-cmd.js
@@ -1,10 +1,9 @@
 const { log } = require('proc-log')
+const BaseCommand = require('./base-cmd.js')
 
 // This is the base for all commands whose execWorkspaces just gets
 // a list of workspace names and passes it on to new Arborist() to
 // be able to run a filtered Arborist.reify() at some point.
-
-const BaseCommand = require('./base-command.js')
 class ArboristCmd extends BaseCommand {
   get isArboristCmd () {
     return true
