+++ lib/base-cmd.js
