+++ lib/commands/access.js
@@ -5,7 +5,7 @@ const pkgJson = require('@npmcli/package-json')
 const localeCompare = require('@isaacs/string-locale-compare')('en')
 const otplease = require('../utils/otplease.js')
 const getIdentity = require('../utils/get-identity.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 const commands = [
   'get',
