+++ lib/commands/cache.js
@@ -3,7 +3,7 @@ const pacote = require('pacote')
 const fs = require('fs/promises')
 const { join } = require('path')
 const semver = require('semver')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 const npa = require('npm-package-arg')
 const jsonParse = require('json-parse-even-better-errors')
 const localeCompare = require('@isaacs/string-locale-compare')('en')
