+++ lib/commands/completion.js
@@ -36,7 +36,7 @@ const Npm = require('../npm.js')
 const { definitions, shorthands } = require('@npmcli/config/lib/definitions')
 const { commands, aliases, deref } = require('../utils/cmd-list.js')
 const { isWindowsShell } = require('../utils/is-windows.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 const fileExists = (file) => fs.stat(file).then(s => s.isFile()).catch(() => false)
