+++ lib/commands/diff.js
@@ -6,7 +6,7 @@ const pacote = require('pacote')
 const pickManifest = require('npm-pick-manifest')
 const { log, output } = require('proc-log')
 const pkgJson = require('@npmcli/package-json')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Diff extends BaseCommand {
   static description = 'The registry diff command'
