+++ lib/commands/dist-tag.js
@@ -4,7 +4,7 @@ const semver = require('semver')
 const { log, output } = require('proc-log')
 const otplease = require('../utils/otplease.js')
 const pkgJson = require('@npmcli/package-json')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class DistTag extends BaseCommand {
   static description = 'Modify package distribution tags'
