+++ lib/commands/doctor.js
@@ -8,7 +8,7 @@ const semver = require('semver')
 const { log, output } = require('proc-log')
 const ping = require('../utils/ping.js')
 const { defaults } = require('@npmcli/config/lib/definitions')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 const maskLabel = mask => {
   const label = []
