+++ lib/commands/edit.js
@@ -2,7 +2,7 @@ const { resolve } = require('path')
 const { lstat } = require('fs/promises')
 const cp = require('child_process')
 const completion = require('../utils/installed-shallow.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 const splitPackageNames = (path) => path.split('/')
 // combine scoped parts
