+++ lib/commands/exec.js
@@ -1,6 +1,6 @@
 const { resolve } = require('path')
 const libexec = require('libnpmexec')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Exec extends BaseCommand {
   static description = 'Run a command from a local or remote npm package'
