+++ lib/commands/explore.js
@@ -3,7 +3,7 @@ const runScript = require('@npmcli/run-script')
 const { join, relative } = require('path')
 const { log, output } = require('proc-log')
 const completion = require('../utils/installed-shallow.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 // npm explore <pkg>[@<version>]
 // open a subshell to the package folder.
