+++ lib/commands/get.js
@@ -1,5 +1,5 @@
 const Npm = require('../npm.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Get extends BaseCommand {
   static description = 'Get a value from the npm configuration'
