+++ lib/commands/help-search.js
@@ -2,7 +2,7 @@ const { readFile } = require('node:fs/promises')
 const path = require('node:path')
 const { glob } = require('glob')
 const { output } = require('proc-log')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 const globify = pattern => pattern.split('\\').join('/')
