+++ lib/commands/help.js
@@ -5,7 +5,7 @@ const { glob } = require('glob')
 const { output } = require('proc-log')
 const localeCompare = require('@isaacs/string-locale-compare')('en')
 const { deref } = require('../utils/cmd-list.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 const globify = pattern => pattern.split('\\').join('/')
