+++ lib/commands/hook.js
@@ -2,7 +2,7 @@ const hookApi = require('libnpmhook')
 const otplease = require('../utils/otplease.js')
 const relativeDate = require('tiny-relative-date')
 const { output } = require('proc-log')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Hook extends BaseCommand {
   static description = 'Manage registry hooks'
