+++ lib/commands/init.js
@@ -8,7 +8,7 @@ const mapWorkspaces = require('@npmcli/map-workspaces')
 const PackageJson = require('@npmcli/package-json')
 const { log, output } = require('proc-log')
 const updateWorkspaces = require('../utils/update-workspaces.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 const posixPath = p => p.split('\\').join('/')
