+++ lib/commands/login.js
@@ -1,7 +1,7 @@
 const { log, output } = require('proc-log')
 const { redactLog: replaceInfo } = require('@npmcli/redact')
 const auth = require('../utils/auth.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Login extends BaseCommand {
   static description = 'Login to a registry user account'
