+++ lib/commands/logout.js
@@ -1,7 +1,7 @@
 const npmFetch = require('npm-registry-fetch')
 const { getAuth } = npmFetch
 const { log } = require('proc-log')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Logout extends BaseCommand {
   static description = 'Log out of the registry'
