+++ lib/commands/org.js
@@ -1,6 +1,6 @@
 const liborg = require('libnpmorg')
 const otplease = require('../utils/otplease.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 const { output } = require('proc-log')
 
 class Org extends BaseCommand {
