+++ lib/commands/owner.js
@@ -4,7 +4,7 @@ const pacote = require('pacote')
 const { log, output } = require('proc-log')
 const otplease = require('../utils/otplease.js')
 const pkgJson = require('@npmcli/package-json')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 const { redact } = require('@npmcli/redact')
 
 const readJson = async (path) => {
