+++ lib/commands/pack.js
@@ -3,7 +3,7 @@ const libpack = require('libnpmpack')
 const npa = require('npm-package-arg')
 const { log, output } = require('proc-log')
 const { getContents, logTar } = require('../utils/tar.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Pack extends BaseCommand {
   static description = 'Create a tarball from a package'
