+++ lib/commands/ping.js
@@ -1,7 +1,7 @@
 const { redact } = require('@npmcli/redact')
 const { log, output } = require('proc-log')
 const pingUtil = require('../utils/ping.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Ping extends BaseCommand {
   static description = 'Ping npm registry'
