+++ lib/commands/pkg.js
@@ -1,6 +1,6 @@
 const { output } = require('proc-log')
 const PackageJson = require('@npmcli/package-json')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 const Queryable = require('../utils/queryable.js')
 
 class Pkg extends BaseCommand {
