+++ lib/commands/profile.js
@@ -5,7 +5,7 @@ const npmProfile = require('npm-profile')
 const qrcodeTerminal = require('qrcode-terminal')
 const otplease = require('../utils/otplease.js')
 const readUserInfo = require('../utils/read-user-info.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 const qrcode = url =>
   new Promise((resolve) => qrcodeTerminal.generate(url, resolve))
