+++ lib/commands/publish.js
@@ -15,7 +15,7 @@ const { getContents, logTar } = require('../utils/tar.js')
 // change that ought to get a RFC written on it.
 const { flatten } = require('@npmcli/config/lib/definitions')
 const pkgJson = require('@npmcli/package-json')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Publish extends BaseCommand {
   static description = 'Publish a package'
