+++ lib/commands/query.js
@@ -1,5 +1,5 @@
 const { resolve } = require('node:path')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 const { log, output } = require('proc-log')
 
 class QuerySelectorItem {
