+++ lib/commands/root.js
@@ -1,5 +1,5 @@
 const { output } = require('proc-log')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Root extends BaseCommand {
   static description = 'Display npm root'
