+++ lib/commands/run-script.js
@@ -1,6 +1,6 @@
 const { log, output } = require('proc-log')
 const pkgJson = require('@npmcli/package-json')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class RunScript extends BaseCommand {
   static description = 'Run arbitrary package scripts'
