+++ lib/commands/sbom.js
@@ -1,5 +1,5 @@
 const localeCompare = require('@isaacs/string-locale-compare')('en')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 const { log, output } = require('proc-log')
 const { cyclonedxOutput } = require('../utils/sbom-cyclonedx.js')
 const { spdxOutput } = require('../utils/sbom-spdx.js')
