+++ lib/commands/search.js
@@ -2,8 +2,8 @@ const Pipeline = require('minipass-pipeline')
 const libSearch = require('libnpmsearch')
 const { log, output } = require('proc-log')
 const formatSearchStream = require('../utils/format-search-stream.js')
+const BaseCommand = require('../base-cmd.js')
 
-const BaseCommand = require('../base-command.js')
 class Search extends BaseCommand {
   static description = 'Search for packages'
   static name = 'search'
