+++ lib/commands/set.js
@@ -1,5 +1,5 @@
 const Npm = require('../npm.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Set extends BaseCommand {
   static description = 'Set a value in the npm configuration'
