+++ lib/commands/shrinkwrap.js
@@ -1,7 +1,7 @@
 const { resolve, basename } = require('path')
 const { unlink } = require('fs/promises')
 const { log } = require('proc-log')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Shrinkwrap extends BaseCommand {
   static description = 'Lock down dependency versions for publication'
