+++ lib/commands/star.js
@@ -2,7 +2,7 @@ const fetch = require('npm-registry-fetch')
 const npa = require('npm-package-arg')
 const { log, output } = require('proc-log')
 const getIdentity = require('../utils/get-identity')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Star extends BaseCommand {
   static description = 'Mark your favorite packages'
