+++ lib/commands/stars.js
@@ -1,7 +1,7 @@
 const fetch = require('npm-registry-fetch')
 const { log, output } = require('proc-log')
 const getIdentity = require('../utils/get-identity.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Stars extends BaseCommand {
   static description = 'View packages marked as favorites'
