+++ lib/commands/team.js
@@ -3,7 +3,7 @@ const libteam = require('libnpmteam')
 const { output } = require('proc-log')
 const otplease = require('../utils/otplease.js')
 
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 class Team extends BaseCommand {
   static description = 'Manage organization teams and team memberships'
   static name = 'team'
