+++ lib/commands/token.js
@@ -2,7 +2,7 @@ const { log, output } = require('proc-log')
 const profile = require('npm-profile')
 const otplease = require('../utils/otplease.js')
 const readUserInfo = require('../utils/read-user-info.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Token extends BaseCommand {
   static description = 'Manage your authentication tokens'
