+++ lib/commands/version.js
@@ -1,7 +1,7 @@
 const { resolve } = require('node:path')
 const { readFile } = require('node:fs/promises')
 const { output } = require('proc-log')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Version extends BaseCommand {
   static description = 'Bump a package version'
