+++ lib/commands/view.js
@@ -10,7 +10,7 @@ const semver = require('semver')
 const { inspect } = require('util')
 const { packument } = require('pacote')
 const Queryable = require('../utils/queryable.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 const readJson = async file => jsonParse(await readFile(file, 'utf8'))
