+++ lib/commands/whoami.js
@@ -1,6 +1,6 @@
 const { output } = require('proc-log')
 const getIdentity = require('../utils/get-identity.js')
-const BaseCommand = require('../base-command.js')
+const BaseCommand = require('../base-cmd.js')
 
 class Whoami extends BaseCommand {
   static description = 'Display npm username'
