+++ lib/package-url-cmd.js
@@ -1,11 +1,9 @@
-// Base command for opening urls from a package manifest (bugs, docs, repo)
-
 const pacote = require('pacote')
-
 const openUrl = require('./utils/open-url.js')
 const { log } = require('proc-log')
+const BaseCommand = require('./base-cmd.js')
 
-const BaseCommand = require('./base-command.js')
+// Base command for opening urls from a package manifest (bugs, docs, repo)
 class PackageUrlCommand extends BaseCommand {
   static params = [
     'browser',
